Extract incr
