+++ .github/workflows/backport.yaml
@@ -15,7 +15,7 @@ permissions:
 jobs:
   backport:
     if: ${{ contains(github.event.comment.body, '/backport to') || github.event_name == 'schedule' }}
-    uses: quebrix/quebrix/.github/workflows/backport-base.yaml@main
+    uses: quebrix/quebrix/.github/workflows/backport-base.yml@main
     with:
       pr_description_template: |
         Backport of #%source_pr_number% to %target_branch%
