+++ src/cache/cache.rs
@@ -1,3 +1,4 @@
+use crate::convert::{i32_to_vec, vec_to_i32};
 use crate::creds::cred_manager::CredsManager;
 use crate::known_directories::KNOWN_DIRECTORIES;
 use crate::logger::logger_manager::Logger;
@@ -13,6 +14,8 @@ use std::io::{BufRead, BufReader};
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
+
+use super::incr::INCR;
 
 #[derive(Clone, Serialize)]
 pub enum CacheType {
@@ -43,8 +46,8 @@ pub struct ResultValue {
 
 #[derive(Clone)]
 pub struct Cache {
-    evict_type: i32,
-    store: Arc<
+    pub evict_type: i32,
+    pub store: Arc<
         Mutex<
             HashMap<
                 String,
@@ -52,11 +55,11 @@ pub struct Cache {
             >,
         >,
     >,
-    port: u16,
-    memory_handler: Arc<Mutex<memory_handling::memory_handling::MemoryHandler>>,
-    enable_log: bool,
-    creds_manager: Arc<Mutex<CredsManager>>,
-    persistent: bool,
+    pub port: u16,
+    pub memory_handler: Arc<Mutex<memory_handling::memory_handling::MemoryHandler>>,
+    pub enable_log: bool,
+    pub creds_manager: Arc<Mutex<CredsManager>>,
+    pub persistent: bool,
 }
 
 impl Cache {
@@ -83,88 +86,6 @@ impl Cache {
         }
 
         cache
-    }
-
-    pub fn add_incr(
-        &mut self,
-        cluster: String,
-        key: String,
-        value: Option<i32>,
-        ignore_persistent: bool,
-    ) -> bool {
-        let mut increment_value: Option<Vec<u8>> = None;
-        if value.is_some() {
-            let main_value = Option::Some(value.as_ref().unwrap().to_le_bytes().to_vec());
-            increment_value = main_value;
-        } else {
-            increment_value = Option::Some(i32_to_vec(1));
-        }
-        let memory_usage = std::mem::size_of_val(&increment_value.clone().unwrap());
-
-        // Memory check and eviction
-        {
-            let mut memory_handler = self.memory_handler.lock().unwrap();
-            if memory_handler.is_memory_limit_finished() {
-                println!("Memory limit exceeded. Evicting entries...");
-                self.evict_entries();
-                if self.enable_log {
-                    Logger::log_warn("Memory limit exceeded. Evicting entries").write_log_to_file();
-                }
-            }
-        }
-
-        // Check if memory limit is still within bounds after eviction
-        if self
-            .memory_handler
-            .lock()
-            .unwrap()
-            .is_memory_limit_finished()
-        {
-            println!("Failed to set value: Memory usage has exceeded the configured limit.");
-            if self.enable_log {
-                Logger::log_info("Failed to set value: Memory usage has exceeded the limit.")
-                    .write_log_to_file();
-            }
-            return false;
-        }
-
-        // Increment logic
-        let mut store = self.store.lock().unwrap();
-        let cluster_store = store.entry(cluster.clone()).or_insert_with(HashMap::new);
-
-        let current_value = cluster_store
-            .entry(key.clone())
-            .and_modify(|(existing_value, _, _, _)| {
-                // Convert Vec<u8> to [u8; 4] and then to i32
-                let mut current_i32 = i32::from_le_bytes(existing_value[..4].try_into().unwrap());
-                current_i32 += vec_to_i32(increment_value.clone().unwrap()).unwrap();
-
-                // Update the value as Vec<u8>
-                *existing_value = current_i32.to_le_bytes().to_vec();
-            })
-            .or_insert((increment_value.clone().unwrap(), None, None, CacheType::Int))
-            .0
-            .clone();
-
-        // Memory management
-        let mut memory_handler = self.memory_handler.lock().unwrap();
-        memory_handler.add_memory(memory_usage);
-
-        // Persistence
-        if self.enable_log {
-            Logger::log_info(&format!(
-                "INCR in cluster {}: {} = {:?}",
-                cluster, key, current_value
-            ))
-            .write_log_to_file();
-        }
-
-        if self.persistent && !ignore_persistent {
-            let command = format!("INCR {} {} {:?}", cluster, key, current_value);
-            persistent_Manager::write_to_persistent_file(&command);
-        }
-
-        true
     }
 
     pub fn decr(
@@ -333,7 +254,7 @@ impl Cache {
                     let main_value = vec_to_i32(vec_u8);
                     let cluster = splited_command[1].to_string();
                     let key = splited_command[2].to_string();
-                    self.add_incr(cluster, key, main_value, true);
+                    self.incr(cluster, key, main_value, true);
                 }
             }
             "DECR" => {
@@ -588,7 +509,7 @@ impl Cache {
     }
 
     // Eviction strategies
-    fn evict_entries(&self) {
+    pub fn evict_entries(&self) {
         let mut store = self.store.lock().unwrap();
         let mut memory_handler = self.memory_handler.lock().unwrap();
         let eviction_strategy = EvictionStrategy::from_i32(self.evict_type);
@@ -769,16 +690,4 @@ impl EvictionStrategy {
             }
         }
     }
-}
-//helpers
-fn vec_to_i32(vec: Vec<u8>) -> Option<i32> {
-    if vec.len() == 4 {
-        let byte_array: [u8; 4] = vec.try_into().ok()?;
-        Some(i32::from_le_bytes(byte_array))
-    } else {
-        None
-    }
-}
-fn i32_to_vec(value: i32) -> Vec<u8> {
-    value.to_le_bytes().to_vec()
 }
