+++ src/cache/mod.rs
@@ -1,4 +1,3 @@
-
 pub mod cache;
-
+pub mod incr;
 pub use cache::Cache;
